Ignore diacritic marks when comparing titles
ISO 4 wants diacritics ignored for comparison, which we have not
done historically.  However, there are some obvious problems with
continuing to ignore this if you try to update to a more recent
version of the ISSN LTWA.

In particular, 'biolog-' is now in the database as 'bìolog-'
(the "i" now has a grave accent in the matching word).
So, failing to treat this properly would break all of biology.